Update firestore.gd
The signal from firestore_task.gd "result_query" sends bod which is an array, not a dictionary. It throws an error because _on_result_query expects a dictionary but works anyways. This fixes the error. Also I have never tried to propose a change hope I am doing this right.